Add files via upload
This is an interim csv file produced in an earlier week and shared among team members. This file is used by the code for data transformation included for this week. The data transformation .ipynb file in the W9 deliverables produces a new csv with no categorical entries -- 'NoCats' - that will be used in week 10 and later.